Renamed chromium-gn.inc to chromium-upstream-tarball.inc
The chromium-gn.inc contained basically the SRC_URI for retrieving
upstream official releases, and the licenses data. So renamed to
have a more accurate name.

Also moved the name and description to chromium.inc, so it can
be included from all Chromium packages, independently of the
actual source package being used.