Decouple X11 build from Chromium upstream version recipes.
Till now, the upstream chromium recipes were only for X11. We are
going to add support to build the upstream ozone wayland from
upstream tarball.

To  do that, first, we need to decouple the recipes of the specific
versions from the X11 build. This is done creating chromium-stable.inc,
chromium-beta.inc and chromium-dev.inc with the version and patches
information. Then chromium-x11-*.bb will just include them and the
chromium-x11.inc build stuff.