Add comment about requirement to still build with python2.
In hardknott, just adding the python2 meta layer is not enough to allow
using it. Configuration should include this variable declaration:
I_SWEAR_TO_MIGRATE_TO_PYTHON3 = "yes"

We are fine with this because we are committed to drop Python2. But
Chromium did not finish migration yet.

So, for hardknott, that variable needs to be set still.